Add generic wrapper + refactor CouncilOverview
