+++ packages/ui/src/overview/components/CouncilOverview/CouncilOverview.tsx
@@ -1,24 +1,19 @@
 import React, { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useHistory } from 'react-router-dom'
-import styled from 'styled-components'
 
-import { ArrowRightLinkIcon } from '@/common/components/icons'
 import { formatDuration } from '@/common/components/statistics/BlockDurationStatistics'
-import { TextMedium } from '@/common/components/typography'
 import { DurationValue } from '@/common/components/typography/DurationValue'
-import { BorderRad, Colors } from '@/common/constants'
 import { CouncilRoutes } from '@/council/constants'
 import { useCouncilConstants } from '@/council/hooks/useCouncilConstants'
 import { useCurrentElection } from '@/council/hooks/useCurrentElection'
 import { useElectionStage } from '@/council/hooks/useElectionStage'
+import { OverviewInfoElement } from '@/overview/components/OverviewInfoElement'
+import { OverviewWrapper } from '@/overview/components/OverviewWrapper'
 
 import { CouncilAnnouncingTiles, CouncilNormalTiles, CouncilRevealingTiles } from './CouncilTilesList'
-import { OverviewInfoElement } from './OverviewInfoElement'
 
 export const CouncilOverview = () => {
   const { t } = useTranslation('overview')
-  const history = useHistory()
   const constants = useCouncilConstants()
   const { stage } = useElectionStage()
   const { election } = useCurrentElection()
@@ -42,6 +37,18 @@ export const CouncilOverview = () => {
     }
   }, [stage])
 
+  const infoElements = useMemo(
+    () => (
+      <>
+        <OverviewInfoElement value={councilSize} label="Council members" />,
+        <OverviewInfoElement value={stageLabel} label={t('council.stage.label')} />,
+        {typeof round !== 'undefined' && <OverviewInfoElement value={`#${round}`} label={t('council.round')} />},
+        <OverviewInfoElement value={nextElectionIn} label={t('council.nextElectionIn')} />,
+      </>
+    ),
+    [councilSize, stageLabel, round, nextElectionIn]
+  )
+
   const scroller = useMemo(() => {
     switch (stage) {
       case 'inactive':
@@ -56,48 +63,11 @@ export const CouncilOverview = () => {
   }, [stage, election])
 
   return (
-    <Wrapper>
-      <Upper>
-        <Title black bold>
-          {t('council.title')}
-        </Title>
-        <ArrowRight onClick={() => history.push(CouncilRoutes.council)} />
-        <InfoElementsWrapper>
-          <OverviewInfoElement value={councilSize} label="Council members" />
-          <OverviewInfoElement value={stageLabel} label={t('council.stage.label')} />
-          {typeof round !== 'undefined' && <OverviewInfoElement value={`#${round}`} label={t('council.round')} />}
-          <OverviewInfoElement value={nextElectionIn} label={t('council.nextElectionIn')} />
-        </InfoElementsWrapper>
-      </Upper>
-      {scroller}
-    </Wrapper>
+    <OverviewWrapper
+      title={t('council.title')}
+      linkPath={CouncilRoutes.council}
+      infoElements={infoElements}
+      scroller={scroller}
+    />
   )
 }
-
-const Wrapper = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  border: 1px solid ${Colors.Black[200]};
-  border-radius: ${BorderRad.s};
-`
-
-const Upper = styled.div`
-  padding: 18px 18px 0;
-`
-
-const Title = styled(TextMedium)`
-  margin-bottom: 25px;
-`
-
-const ArrowRight = styled(ArrowRightLinkIcon)`
-  position: absolute;
-  top: 24px;
-  right: 18px;
-  cursor: pointer;
-`
-
-const InfoElementsWrapper = styled.div`
-  display: flex;
-  column-gap: 50px;
-`
